Fix wrong output data json format when compiling with GCC
This change works around an issue when compiling with GCC where by
printing only a '\n' the produced global output is wrong. What happens
is that new lines are actually printed before the content in the
precedent fprintf calls. Hence the generated json is unparsable.

Note also that replacing the LOG macros with direct calls to fprintf
does _not_ solve the issue.

Change-Id: I2d97703cbe7703aae94def1943e2ac693945ab31