Add size of data in bytes to output tensor print
The printing of the output tensor can be truncated, by adding the
original size of the output in bytes the included data can be
checked for completeness. The dims will tell the number of elements
of the output tensor and is not useful for this purpose as the
element size can vary.

Change-Id: Ied5447e34416a8c6668eeae5c01a5474f103023f