Refactor performance measurements
Change 'Inference runtime' to measure CPU cycles for the
Tensorflow Lite Micro interpreter.Invoke() call.

Add 'Operator(s) runtime' print that prints a summary for
cycles spent on all operators during an inference. (This is
equivalent to the old reported 'Inference runtime')

Move prints out of the EndEvent() function in ArmProfiler as
it otherwise interferes with the inference cycle measurement.

Change-Id: Ie11b5abb5b12a3bcf5a67841f04834d05dfd796d