Remove CMake workaround for Arm Clang
Previous version of CMake did not allow CMAKE_SYSTEM_PROCESSOR
to contain CPU features for Arm Clang toolchain. A workaround
was added to manually strip the CPU features.

Recent versions of CMake have removed this limitation and instead
requires the toolchain to manually add -mcpu=<CPU+features>, which
means that the workaround can be removed.

Change-Id: I3e8647c47df27a96cdd103614ecf2540f150ab36