Fix parser not always null-terminating model desc
The inference parser copies the description from the model using
strncpy, which will only add a null-terminator if the source string is
shorter than the destination. Otherwise, if the bytes copied are not
null-terminated, the destination is left without one.

To ensure that the description is always null-terminated, the last byte
of the description is now always overwritten with one.

Change-Id: I49c23acd12d661e1f5c37088dba2e37935f25fa5
Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>